scripts/inference/onnx/bounding_box/requirements.txt
tensorflow==2.3.0
numpy==1.18.5
Pillow==7.2.0
matplotlib==3.1.3
opencv-python==4.5.1.48
requests==2.25.1
datature-hub==0.2.0
scipy==1.4.1
onnxruntime==1.12.1
